+++ src/CronScheduler/Cron/ICrontabField.cs
@@ -1,9 +1,0 @@
-﻿namespace CronScheduler.AspNetCore.Cron
-{
-    public interface ICrontabField
-    {
-        int GetFirst();
-        int Next(int start);
-        bool Contains(int value);
-    }
-}