+++ src/CronScheduler/SchedulerBuilder.cs
@@ -6,6 +6,9 @@ namespace CronScheduler.AspNetCore
 {
     public class SchedulerBuilder
     {
+        /// <summary>
+        /// EventHanlder for Startup for Hosted Apps.
+        /// </summary>
         public EventHandler<UnobservedTaskExceptionEventArgs> UnobservedTaskExceptionHandler = null;
 
         public SchedulerBuilder(IServiceCollection services)
@@ -13,6 +16,9 @@ namespace CronScheduler.AspNetCore
             Services = services;
         }
 
+        /// <summary>
+        /// <see cref="IServiceCollection"/> for the DI.
+        /// </summary>
         public IServiceCollection Services { get; }
     }
 }