+++ src/CronScheduler/SchedulerTaskWrapper.cs
@@ -1,20 +1,30 @@
 ﻿using System;
-using CronScheduler.AspNetCore.Cron;
+using Cronos;
 
 namespace CronScheduler.AspNetCore
 {
     internal class SchedulerTaskWrapper
     {
-        public CrontabSchedule Schedule { get; set; }
-        public IScheduledJob Task { get; set; }
+        public CronExpression Schedule { get; set; }
+        public IScheduledJob ScheduledJob { get; set; }
 
         public DateTime LastRunTime { get; set; }
         public DateTime NextRunTime { get; set; }
 
+        public SchedulerTaskWrapper(
+            CronExpression cronExpression,
+            IScheduledJob scheduledJob,
+            DateTime nextRunTime)
+        {
+            Schedule = cronExpression;
+            ScheduledJob = scheduledJob;
+            NextRunTime = nextRunTime;
+        }
+
         public void Increment()
         {
             LastRunTime = NextRunTime;
-            NextRunTime = Schedule.GetNextOccurrence(NextRunTime);
+            NextRunTime = Schedule.GetNextOccurrence(NextRunTime).Value;
         }
 
         public bool ShouldRun(DateTime currentTime)