+++ src/CronSchedulerApp/Program.cs
@@ -5,7 +5,9 @@ using Microsoft.Extensions.Logging;
 
 namespace CronSchedulerApp
 {
+#pragma warning disable RCS1102 // Make class static.
     public class Program
+#pragma warning restore RCS1102 // Make class static.
     {
         public static void Main(string[] args)
         {