+++ CHANGELOG.md
@@ -8,6 +8,13 @@ Version 1.0.6 (2018-01-14)
 
 * Resovled issue #6 "Add support for kdcllc Docker image" 
 
+* Resolved issue #4 "Add support for seconds with Cron"
+
+```c#
+   public string CronTimeZone { get; };
+
+```
+
 Version 1.0.5 (2018-03-12)
 ----------------------------
  * Resolved issue#1 "Add option to not run job on application start"